Added yet another dirty hack
Paramiko dynamically imports libraries during threading mostly
in the encoding and cryptography libraries.  If SSH code is
executing during an import the main thread will have a lock
on import modules and when parimiko tries to do a dynamic import
it will hang indefinitely blocking both threads.  In order to
prevent this when I encounter such a hang I add the call that
is doing the dynamic import to the main client.  This preimports
it before threading begins and allows the thread to used the
already imported copy. It's a neverending battle but it makes
sshaolin more stable when daring to run SSH code during an import.
If you see a problem like this in your application the way to
find the problem call is to run your code like so.
python -m trace --trace some_code.py > output.trace
This will create a very large debug trace. When the following
starts repeating look right above for the problem call.
threading.py(351):                 while True:
threading.py(352):                     gotit = waiter.acquire(0)
threading.py(353):                     if gotit:
threading.py(355):                     remaining = endtime - _time()